deleted:    Untitled-1.py 	new file:   __pycache__/add_pages.cpython-311.pyc 	modified:   __pycache__/add_pages.cpython-39.pyc 	new file:   __pycache__/create_par.cpython-311.pyc 	modified:   __pycache__/create_par.cpython-39.pyc 	new file:   __pycache__/edit_page.cpython-311.pyc 	new file:   __pycache__/edit_page.cpython-39.pyc 	new file:   __pycache__/move_tags.cpython-311.pyc 	new file:   __pycache__/move_tags.cpython-39.pyc 	new file:   __pycache__/select_files.cpython-311.pyc 	modified:   __pycache__/select_files.cpython-39.pyc 	modified:   add_pages.py 	modified:   create_par.py 	modified:   edit_page.py 	new file:   html_files/html/fds.html 	new file:   html_files/html/sdfa.html 	modified:   html_files/index.html 	modified:   html_parser.py 	new file:   move_tags.py 	new file:   saves/selected_id.txt 	modified:   select_files.py

+++ create_par.py
@@ -4,26 +4,53 @@ def create_paragraph():
         def get_input():
             #make the input into a readable string
             input_value = textBox.get("1.0","end-1c")
-            #check if user entered any text
-            if input_value!='':
-                #imbort beautiful soup
-                from bs4 import BeautifulSoup as bs
+            #id_input = id_name.get()
+            id_input = id_textBox.get("1.0", "end-1c")
+            #check if the id already exists
+            #imbort beautiful soup
+            from bs4 import BeautifulSoup as bs
+            #get the html file from the saved file
+            text_file = open("saves/selected_file.txt", "r")
+            get_file = text_file.readline()
+            if get_file == "index.html":
                 with open("html_files/index.html") as fp:
                     soup = bs(fp) 
-                #make a new paragraph tag
+            else:
+                with open("html_files/html"+get_file) as fp:
+                    soup = bs(fp) 
+            divs = soup.find_all('div')
+            id_exists = False
+            #loop throught all divs to see if the id already exists
+            for div in divs:
+                if div['id'] == id_input:
+                    id_exists = True
+                else:
+                    id_exists = False
+            #check if user entered any text
+            if input_value!='' and id_input!='' and id_exists == False:
+
+                #make a new paragraph tag and a div tag
+                new_div = soup.new_tag("div")
                 new_par = soup.new_tag("p")
+                #give the div an id
+                new_div['id'] = id_input
                 #add the string
                 new_par.string = input_value
                 #add the tag after body
-                soup.html.body.append(new_par)
+                soup.html.body.append(new_div)
+                #add the paragraph adter the div
+                new_div.append(new_par)
                 #open file to write
-                html_file = open("html_files/index.html", "w") 
+                if get_file == "index.html":
+                    html_file = open("html_files/index.html", 'w')
+                else:
+                    html_file =  open("html_files/html"+get_file, "w")
                 #write to file
                 html_file.writelines(str(soup))
                 html_file.close()
                 root.destroy()
             else:
-                tk.Label(text = "You must enter some text!").place(x = 40, y = 60) 
+                tk.Label(text = "You must enter some text or a non-existing id!").pack() 
         #import tkinter
         import tkinter as tk
         #new window
@@ -33,6 +60,10 @@ def create_paragraph():
         #new input box
         textBox = tk.Text(root, height = 5, width = 20)
         textBox.pack()
+        #input for id
+        tk.Label(root, text="Enter the id of the paragraph").pack()  
+        id_textBox = tk.Text(root, height = 5, width = 5)
+        id_textBox.pack()
         #new button
         printButton = tk.Button(root, text = "Save", command=get_input)
         printButton.pack()
